Move helper function to init
